refactor model file, the great union
